+++ resources/js/components/ExceptionCodePreview.vue
@@ -22,6 +22,7 @@
 <template>
     <pre class="code-bg px-4 mb-0 text-white">
         <p v-for="(line, number) in lines"
+           :key="number"
            class="mb-0"
            :class="{'highlight': number == highlightedLine}"><span class="mr-4">{{number}}</span> <span v-html="highlight(line, number)" /></p>
     </pre>