+++ src/ExtractsMailableTags.php
@@ -14,12 +14,16 @@ trait ExtractsMailableTags
      */
     protected static function registerMailableTagExtractor()
     {
-        Mailable::buildViewDataUsing(function ($mailable) {
-            return [
+        $existingCallback = Mailable::$viewDataCallback;
+
+        Mailable::buildViewDataUsing(function ($mailable) use ($existingCallback) {
+            $existingData = $existingCallback ? call_user_func($existingCallback, $mailable) : [];
+
+            return array_merge($existingData, [
                 '__telescope' => ExtractTags::from($mailable),
                 '__telescope_mailable' => get_class($mailable),
                 '__telescope_queued' => in_array(ShouldQueue::class, class_implements($mailable)),
-            ];
+            ]);
         });
     }
 }